Reorganization of the project files.
In particular, the board project files are now generated automatically, to
ease the addition of new boards and make sure the projects are homogeneous.

Also the handling of the common project is now simplified and shared among
all boards.

Finally, the runtime used is now supposed to be installed in the standard
RTS search path, so this patch removes the explicit dependency over
the embedded-runtimes repository.